requirements.txt
# Core ML
torch>=2.2.0
torchvision>=0.17.0

# HuggingFace
transformers>=4.40.0
datasets>=2.19.0
tokenizers>=0.19.0
accelerate>=0.29.0

# scikit-learn (baseline)
scikit-learn>=1.4.0
joblib>=1.3.0

# Data & plotting
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.2.0

# Config
PyYAML>=6.0.1

# Progress bars
tqdm>=4.66.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
